Move helper methods to ReviewingTestCase.
